pyproject.toml
[tool.poetry]
name = "nonebot-plugin-ba-tools"
version = "0.4.4"
description = "一个BlueArchive的工具箱"
authors = [
  "hanasaki <hanasakayui2022@gmail.com>",
  "kawaiior <kawaiiorv@gmail.com>",
]
license = "MIT"
readme = "README.md"
packages = [{ include = "nonebot_plugin_ba_tools" }]
homepage = "https://github.com/hanasa2023/nonebot-plugin-ba-tools#readme"
repository = "https://github.com/hanasa2023/nonebot-plugin-ba-tools"

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
httpx = ">=0.27.0, <0.28.0"
nonebot2 = ">=2.3.2"
pydantic = ">=1.0, <3.0.0"
beautifulsoup4 = ">=4.12.3"
nonebot-adapter-onebot = ">=2.1.0"
nonebot-plugin-alconna = ">=0.52.2"
nonebot-plugin-localstore = ">=0.7.1"
nonebot-plugin-apscheduler = ">=0.5.0"
nonebot-plugin-htmlrender = ">=0.3.3"
pillow = ">=10.4.0"
nonebot-plugin-uninfo = "^0.5.0"
tortoise = "^0.1.1"
openai = "^1.60.2"
pyyaml = "^6.0.2"
deepdiff = "^8.1.1"

[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.ustc.edu.cn/pypi/simple/"
priority = "primary"


[tool.poetry.group.test.dependencies]
nonebug = "^0.3.7"
pytest-asyncio = "^0.23.8"

[tool.poetry.group.dev.dependencies]
mypy = ">=1.11.1"
nonebot2 = { version = ">=2.3.2", extras = ["fastapi", "uvicorn"] }
nonebot-adapter-onebot = ">=2.1.0"
basedpyright = ">=1.16.0"
ruff = "^0.6.1"
types-beautifulsoup4 = ">=4.12.0.20240511"
datamodel-code-generator = ">=0.26.0"
types-aiofiles = "^24.1.0.20240626"
pre-commit = "^3.8.0"
types-pyyaml = "^6.0.12.20241230"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "F401"]

[tool.basedpyright]
typeCheckingMode = "standard"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
adapters = [
  { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = []
plugin_dirs = [""]
builtin_plugins = ["echo", "single_session"]
